plz no crash

+++ 2019/s1/adds/assignment8/BubbleSort.cpp
@@ -20,7 +20,7 @@ void BubbleSort::doSort() {
     {
         sorted = true;
 
-        for(int i = 0; i < list.size() - 1; i++)
+        for(int i = 0; i < (signed int)(list.size() - 1); i++)
         {
             if (list[i] > list[i+1])
             {
